backend/requirements.txt
# brotlipy==0.7.0
# certifi==2023.7.22
# cffi==1.15.1
# charset-normalizer==2.0.4
# click==8.1.8
# psycopg2-binary==2.9.10
# pycparser==2.21
# pyOpenSSL==22.0.0
# PySocks==1.7.1
# requests==2.28.1
# six==1.16.0
# toolz==0.12.0
# tqdm==4.64.1
# typing_extensions==4.13.2
# urllib3==1.26.13
# uvicorn==0.34.2

# # Flask and related packages
# Flask==2.2.3
# Flask-Cors==3.0.10
# Flask-RESTful==0.3.9
# Flask-SQLAlchemy==3.0.3
# SQLAlchemy==2.0.4

# # Other dependencies your app might need
# python-dotenv==1.0.0
# PyYAML==6.0.1  # Replacement for ruamel.yaml
# gunicorn==23.0.0  # For production deployment

Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask_restx
python-dotenv>=1.0.1,<2.0.0
psycopg2-binary==2.9.9
pydantic==2.5.2
transformers==4.36.0
torch==2.1.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
langchain==0.1.0
pytest==7.4.3
black==23.11.0
isort==5.12.0
mypy==1.7.1
# Supabase integration
supabase==2.3.0
# Zilliz Cloud / Milvus integration
pymilvus==2.3.4
langchain-community==0.0.19
sentence-transformers==2.5.1
numpy>=1.24.0
# OpenAI integration
openai==1.3.0
flask_jwt_extended
gunicorn==21.2.0